backend/requirements.txt
# Moomingle Backend - Cattle Breed Classifier API
# For Render deployment with Gunicorn

flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
huggingface-hub>=0.16.0
onnxruntime>=1.15.0
numpy>=1.24.0
Pillow>=10.0.0
torchvision>=0.15.0
torch>=2.0.0

# For Gradio interface (HF Spaces)
gradio>=4.0.0
